+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
-## v0.8.3
+## v0.9.0
 
-- Added z Index property to Overlay and removed from modal
+- A new variant was added on the button, infoLight.
+- IconText Component was added.
+- Adjustment in the ratingBus, ratingPunctuality and ratingStaff icons.
+- Text size in the Rating component.